12/31 상세내용 출력

+++ toy_post2_addDB/src/main/java/com/stone/springmvc/common/Board.java
@@ -3,15 +3,15 @@ package com.stone.springmvc.common;
 import java.util.ArrayList;
 
 public class Board {
-	long no;
+	int no;
 	String title;
 	String contents;
 
-	public long getNo() {
+	public int getNo() {
 		return no;
 	}
 
-	public void setNo(long no) {
+	public void setNo(int no) {
 		this.no = no;
 	}
